Applying CR changes and fixing first capital letter
